+++ delete.go
@@ -9,10 +9,13 @@ func syncdel(files *[]string) {
 	for _, path := range *files {
 		currentTaskType = 4
 		currentFile = path
-		err = os.RemoveAll(path)
 
-		if err != nil {
-			errDeletingFile(path, err)
+		if !dryrun {
+			err = os.RemoveAll(path)
+
+			if err != nil {
+				errDeletingFile(path, err)
+			}
 		}
 	}
 }
